SURFACE: corrections in dxf_surface_extruded_*_class_ID().

+++ src/surface.h
@@ -425,8 +425,8 @@ double dxf_surface_extruded_get_align_angle (DxfSurfaceExtruded *extruded_surfac
 DxfSurfaceExtruded *dxf_surface_extruded_set_align_angle (DxfSurfaceExtruded *extruded_surface, double align_angle);
 int16_t dxf_surface_extruded_get_sweep_alignment_option (DxfSurfaceExtruded *extruded_surface);
 DxfSurfaceExtruded *dxf_surface_extruded_set_sweep_alignment_option (DxfSurfaceExtruded *extruded_surface, int16_t sweep_alignment_option);
-int16_t dxf_surface_extruded_get_class_ID (DxfSurfaceExtruded *extruded_surface);
-DxfSurfaceExtruded *dxf_surface_extruded_set_class_ID (DxfSurfaceExtruded *extruded_surface, int16_t class_ID);
+int32_t dxf_surface_extruded_get_class_ID (DxfSurfaceExtruded *extruded_surface);
+DxfSurfaceExtruded *dxf_surface_extruded_set_class_ID (DxfSurfaceExtruded *extruded_surface, int32_t class_ID);
 
 
 #ifdef __cplusplus
